nir: define behavior of nir_op_bfm and nir_op_u/ibfe according to SM5 spec.
That is: the five least significant bits provide the values of 'bits' and 'offset'
which is the case for all hardware currently supported by NIR and using the bfm/bfe instructions.
This patch also changes the lowering of bitfield_insert/extract using shifts to not use bfm anymore.
Thus, the flag 'lower_bfm is removed'.