amd/compiler: rework phi insertion in the RA
read_variable() has been split and phis are now only created when
initializing the register file from the live-in.

Also, instead of adding the phis to the block at the end, the phis are
added while initializing the register file. This results in them being
handled normally and has their definition registers be decided upon more
robustly.